+++ tests/unit/models/reddit/test_subreddit.py
@@ -1,5 +1,4 @@
 import json
-import os
 import pickle
 import pytest
 from unittest import mock